Enable Distributed Tracing with Tanzu Observability
What if our backend started to be slow? This sample application has a
service that we can use to enable a random latency for some requests.

For instance, this slows down 30% of "some requests" on the backend:

```
http :8081/actuator/latency ratio=0.3
```

Distributed traces are a great way to figure out what is going on a
problematic request. This commit adds the Spring Cloud Sleuth starter
that provides distributed tracing support for Spring applications.

Traces are emitted to a backend and Sleuth supports several of them.
This commit also adds the Tanzu Observability with Wavefront starter
which also emits metrics for our application. By default, the starter
creates a freemium account automatically.

To help identifying our app we add the name of the application and the
name of the service. Because this is a demo, we also increase the
frequency at which traces and metrics are shipped to Wavefront.